Renamed package `logfilter` to `filter`
